docs: main파일 코드를 UserCards, UserCardsError, RockPaperScissors 파일로 분할
