+++ RockPaperScissors/RockPaperScissors/UserCards.swift
@@ -1,0 +1,13 @@
+//
+//  UserCards.swift
+//  RockPaperScissors
+//
+//  Created by 박종화 on 2023/05/03.
+//
+
+enum UserCards: Int {
+    case rock = 1
+    case scissors = 2
+    case paper = 3
+    case exit = 0
+}
