+++ RockPaperScissors/RockPaperScissors/UserCardsError.swift
@@ -1,0 +1,10 @@
+//
+//  UserCardsError.swift
+//  RockPaperScissors
+//
+//  Created by 박종화 on 2023/05/03.
+//
+
+enum UserCardsError: Error {
+    case error
+}
